Successfully separated byte_data from file_dict for all measurement data. adapted multiprocessing to also work with memory mapping. needs testing.

+++ data_analysis/example_scripts/Script for separating byte_data from all measurement files and resaving using Numpy save.py
@@ -8,7 +8,7 @@
 #       format_version: '1.3'
 #       jupytext_version: 1.13.8
 #   kernelspec:
-#     display_name: Python 3 (ipykernel)
+#     display_name: Python 3
 #     language: python
 #     name: python3
 # ---
@@ -18,7 +18,8 @@
 # * DO NOT 'Run All'!
 # * Please run cell-by-cell and read carefully the output of each cell.
 # * Must be run from a PC which has all data on board (not in cloud), i.e. the Lab PC
-# * THIS IS A ONE-TIME SCRIPT - running on already converted data will FAIL
+# * THIS IS A ONE-TIME SCRIPT - but can safely be run on already converted data (KeyError will be caught)
+# * Does not currently seem to work with old .mat files. A fallback was created for them in correlation_function.py
 
 # %%
 from pathlib import Path
@@ -76,14 +77,15 @@ solution_dir_paths = [
 #     if file_path.suffix == ".pkl"
 # ]
 
+print(f"{len(solution_dir_paths)} 'date' folders:")
 for dir_path in solution_dir_paths:
     print(dir_path)
 
 # %%
-print(f"Seperating byte_data from file_dicts... ")
-for solution_dir_path in solution_dir_paths:
+print(f"Seperating byte_data from file_dicts in {len(solution_dir_paths)} folders")
+for idx, solution_dir_path in enumerate(solution_dir_paths):
 
-    print(f"{solution_dir_path}... ", end="")
+    print(f"{idx} - {solution_dir_path}... ", end="")
 
     # get all .pkl files in a 'solution' directory
     file_paths = [
@@ -94,14 +96,51 @@ for solution_dir_path in solution_dir_paths:
     # Numpy-saving byte_data separately with the same name + byte_data.npy
     # removing byte_data key from file_dict, then resaving using save_object (compressed with blosc).
     for file_path in file_paths:
-        file_dict = load_object(file_path)
-        byte_data = file_dict["full_data"].pop("byte_data")
-        np.save(
-            file_path.with_name(file_path.name.replace(".pkl", "_byte_data.npy")),
-            byte_data,
-            allow_pickle=False,
-            fix_imports=False,
-        )
-        save_object(file_dict, file_path, "blosc")
+        converted_file_path = file_path.with_name(file_path.name.replace(".pkl", "_byte_data.npy"))
+        if not converted_file_path.is_file():
+            try:
+                file_dict = load_object(file_path)
+            except Exception as exc:
+                print(
+                    f" Exception [{exc}] raised for file path {file_path}. Skipping, but should be checked out! ",
+                    end="",
+                )
+                continue
+            try:
+                byte_data = file_dict["full_data"].pop("byte_data")
+            except TypeError as exc:
+                # file dict is actually a list??
+                file_dict = file_dict[0]
+                byte_data = file_dict["full_data"].pop("byte_data")
+                np.save(
+                    file_path.with_name(file_path.name.replace(".pkl", "_byte_data.npy")),
+                    byte_data,
+                    allow_pickle=False,
+                    fix_imports=False,
+                )
+                save_object(file_dict, file_path, "blosc")
+                print("V", end="")
+            except KeyError:
+                # data key is "data" and not "byte_data"
+                byte_data = file_dict["full_data"].pop("data")
+                np.save(
+                    file_path.with_name(file_path.name.replace(".pkl", "_byte_data.npy")),
+                    byte_data,
+                    allow_pickle=False,
+                    fix_imports=False,
+                )
+                save_object(file_dict, file_path, "blosc")
+                print("V", end="")
+            else:
+                np.save(
+                    file_path.with_name(file_path.name.replace(".pkl", "_byte_data.npy")),
+                    byte_data,
+                    allow_pickle=False,
+                    fix_imports=False,
+                )
+                save_object(file_dict, file_path, "blosc")
+                print("V", end="")
 
-print("Done.")
+    print(" Done.")
+
+print("\n\nOperation finished!")
